requirement.txt
streamlit
faiss-cpu
sentence-transformers
python-docx
PyPDF2
numpy
